LayoutTests/fast/forms/ios/accessory-bar-navigation-obscured-input-expected.txt

Tests that obscured inputs are not focusable using the accessory bar.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Focus the unobscured input.
PASS document.activeElement is input

Attempting to to use the accessory bar to focus an obscured input should preserve the existing focus.
PASS document.activeElement is input

Attempting to to use the accessory bar to focus an unobscured input should change focus.
PASS document.activeElement is iframe
PASS successfullyParsed is true

TEST COMPLETE
